+++ .github/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### New features since last release
 
+* Add `qml.StatePrep()` and `qml.QubitStateVector()` support to `lightning.tensor`.
+  [(#849)](https://github.com/PennyLaneAI/pennylane-lightning/pull/849)
+
 * Add LightningGPU Linux (aarch64+GraceHopper) wheels to PyPI.
   [(#815)](https://github.com/PennyLaneAI/pennylane-lightning/pull/815)